Sprint/Sprint Implementation/Exam_Office_System/myproject/Answer_Script_Management/test_views.py
import pytest
from unittest.mock import patch, MagicMock
from django.urls import reverse

# Mocking Django components
@pytest.fixture
def mock_user():
    user_mock = MagicMock()
    user_mock.username = 'testuser'
    user_mock.password = 'testpass'
    return user_mock

@pytest.fixture
def mock_exam():
    exam_mock = MagicMock()
    exam_mock.id = 1
    exam_mock.department_id = 1
    exam_mock.batch = '2023'
    exam_mock.session = '2023/2024'
    exam_mock.exam_date = '2023-12-01'
    exam_mock.course_id = 1
    return exam_mock

@pytest.fixture
def mock_student():
    student_mock = MagicMock()
    student_mock.id = 1
    student_mock.registration_number = 'REG123'
    student_mock.department_id = 1
    student_mock.session = '2023/2024'
    student_mock.name = 'Test Student'
    return student_mock

@pytest.fixture
def mock_answer_script(mock_exam, mock_student):
    answer_script_mock = MagicMock()
    answer_script_mock.id = 1
    answer_script_mock.exam = mock_exam
    answer_script_mock.student = mock_student
    return answer_script_mock

@pytest.fixture
def load_test_cases():
    return {
        'manage_answer_scripts': {'GET': {'status_code': 200}},
        'view_attendance_and_scripts': {'GET': {'status_code': 200}},
        'evaluate_script': {'GET': {'status_code': 200}},
        'finalize_management': {'POST': {'status_code': 302, 'redirect_url': '/success/'}}
    }

@patch('Answer_Script_Management.views.User', autospec=True)
def test_manage_answer_scripts(mock_user, load_test_cases):
    mock_user.objects.create_user.return_value = mock_user
    response = MagicMock()  # Mock response object
    response.status_code = load_test_cases['manage_answer_scripts']['GET']['status_code']
    response.context = {'exams': []}  # Mock context

    # Simulate view logic
    assert response.status_code == load_test_cases['manage_answer_scripts']['GET']['status_code']
    assert 'exams' in response.context

@patch('Answer_Script_Management.views.Exam', autospec=True)
@patch('Answer_Script_Management.views.Student', autospec=True)
@patch('Answer_Script_Management.views.AnswerScript', autospec=True)
def test_view_attendance_and_scripts(mock_exam, mock_student, mock_answer_script, load_test_cases):
    response = MagicMock()  # Mock response object
    response.status_code = load_test_cases['view_attendance_and_scripts']['GET']['status_code']
    response.context = {'exam': mock_exam, 'students': [mock_student], 'answer_scripts': [mock_answer_script]}

    # Simulate view logic
    assert response.status_code == load_test_cases['view_attendance_and_scripts']['GET']['status_code']
    assert 'exam' in response.context
    assert 'students' in response.context
    assert 'answer_scripts' in response.context

@patch('Answer_Script_Management.views.AnswerScript', autospec=True)
def test_evaluate_script(mock_answer_script, load_test_cases):
    response = MagicMock()  # Mock response object
    response.status_code = load_test_cases['evaluate_script']['GET']['status_code']
    response.context = {'answer_script': mock_answer_script}

    # Simulate view logic
    assert response.status_code == load_test_cases['evaluate_script']['GET']['status_code']
    assert 'answer_script' in response.context

def test_finalize_management(load_test_cases):
    response = MagicMock()  # Mock response object
    response.status_code = load_test_cases['finalize_management']['POST']['status_code']
    response.url = load_test_cases['finalize_management']['POST']['redirect_url']

    # Simulate view logic
    assert response.status_code == load_test_cases['finalize_management']['POST']['status_code']
    assert response.url == load_test_cases['finalize_management']['POST']['redirect_url']